+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - Parser
   - **Breaking Change**: `ASTNodeQuerySelector` a.k.a the `$` function will return a list of `ASTNode` instead of `ASTNode | undefined`, and can only query by type name not literals.
   - **Breaking Change**: Remove `GrammarRule.queryIndex`.
+  - Add `ParserBuilder.cascadeQueryPrefix` to use cascade query in the `$` function.
   - Add `AdvancedBuilder` to build ELR parser, you can use `|+*()?` in your grammar rules.
 
 ## v0.5.1